requirements.txt
paho-mqtt
psycopg2-binary
Flask
Flask-Cors